requirements.txt
pytest
grpcio
grpcio-tools
fastapi
uvicorn
pydantic
httpx
